move get_plural_of_entity and get_singular_of_entity to utils.py
