+++ merlin/backends/sqlite/sqlite_store_base.py
@@ -28,6 +28,7 @@ from typing import Any, Dict, Generic, List, Optional, Tuple, Type
 from merlin.backends.sqlite.sqlite_connection import SQLiteConnection
 from merlin.backends.store_base import StoreBase, T
 from merlin.backends.utils import deserialize_entity, get_not_found_error_class, serialize_entity
+from merlin.utils import get_plural_of_entity
 
 
 LOG = logging.getLogger(__name__)
@@ -221,8 +222,6 @@ class SQLiteStoreBase(StoreBase[T], Generic[T]):
         Returns:
             A list of matching entities.
         """
-        from merlin.cli.utils import get_plural_of_entity  # pylint: disable=import-outside-toplevel
-
         entity_type = get_plural_of_entity(self.table_name, split_delimiter="_", join_delimiter=" ")
         log_action = "filtered" if filters else "all"
         LOG.info(f"Fetching {log_action} {entity_type} from SQLite{f' with filters: {filters}' if filters else ''}...")
